Integrate port with sidekick module
